@@ -4,7 +4,7 @@
   "description": "Dishin' App",
   "main": "index.js",
   "scripts": {
-    "dev": "node server.js",
+    "dev": "NODE_ENV=development node server.js",
     "prod": "NODE_ENV=production node server.js",
     "build": "next build"
   },
@@ -14,6 +14,7 @@
     "@material-ui/core": "^3.9.2",
     "@material-ui/icons": "^3.0.2",
     "@material-ui/lab": "^3.0.0-alpha.30",
+    "@material-ui/styles": "^4.10.0",
     "@svgr/webpack": "^5.4.0",
     "body-parser": "^1.18.3",
     "classnames": "^2.2.6",
@@ -22,6 +23,8 @@
     "downshift": "^3.2.7",
     "emotion": "^10.0.17",
     "express": "^4.16.4",
+    "geolib": "^3.3.1",
+    "google-map-react": "^2.0.8",
     "i": "^0.3.6",
     "immutable": "^4.0.0-rc.12",
     "isomorphic-fetch": "^2.2.1",
@@ -32,11 +35,12 @@
     "material-ui-chip-input": "^1.0.0-beta.15",
     "material-ui-dropzone": "2.3.5",
     "material-ui-popup-state": "^1.4.1",
+    "material-ui-scrollable-tabs": "^2.0.0",
     "mobx": "^5.9.0",
     "mobx-react": "^5.4.3",
     "moment": "^2.24.0",
     "net": "^1.0.2",
-    "next": "^7.0.2",
+    "next": "^9.5.1",
     "next-images": "^1.0.4",
     "next-redux-saga": "^3.0.0",
     "next-redux-wrapper": "^3.0.0-alpha.0",
@@ -53,8 +57,9 @@
     "react-facebook-share-link": "^1.0.4",
     "react-geocode": "^0.1.2",
     "react-jss": "^8.6.1",
-    "react-redux": "^6.0.0",
+    "react-redux": "^7.2.1",
     "react-router-redux": "^4.0.8",
+    "react-scroll-to": "^3.0.0-beta.6",
     "react-select": "^2.4.1",
     "react-share": "^2.4.0",
     "react-swipeable-views": "^0.13.1",
@@ -66,19 +71,21 @@
     "redux-logger": "^3.0.6",
     "redux-saga": "^1.0.1",
     "redux-thunk": "^2.3.0",
+    "scrollparent": "^2.0.1",
     "slug": "^1.0.0"
   },
   "devDependencies": {
+    "babel-eslint": "^10.1.0",
     "eslint": "^5.14.1",
     "eslint-config-airbnb": "^17.1.0",
     "eslint-config-prettier": "^4.0.0",
     "eslint-plugin-import": "^2.16.0",
     "eslint-plugin-jsx-a11y": "^6.2.1",
     "eslint-plugin-prettier": "^3.0.1",
+    "eslint-plugin-react": "^7.20.5",
     "husky": "^1.3.1",
     "lint-staged": "^8.1.4",
-    "prettier": "^1.16.4",
-    "stylelint": "^9.10.1",
+    "stylelint": "^13.6.1",
     "stylelint-config-prettier": "^4.0.0",
     "stylelint-config-standard": "^18.2.0"
   }
