@@ -1,36 +1,35 @@
-import React, { PureComponent } from 'react';
-import { withStyles } from '@material-ui/core/styles';
-import Grid from '@material-ui/core/Grid';
-import Button from '@material-ui/core/Button';
-import AppLayout from '../components/layouts/AppLayout';
-import { Typography } from '@material-ui/core';
-import styles from '../styles/common';
-import classnames from 'classnames';
-import Link from 'next/link';
-import AutoComplete from '../components/autoComplete';
-import notify from '../utils/notifier';
-import { userAPI } from '../services/userAPI';
-import Geocode from 'react-geocode';
-import { APP_URL, API_IMAGE_URL } from '../utils/config';
-import RoomIcon from '@material-ui/icons/Room';
-import { Scrollbars } from 'react-custom-scrollbars';
-import SocialLinks from '../components/common/SocialLinks';
-import actions from '../redux/global/actions';
-import { connect } from 'react-redux';
-import WindowResizeListener from 'react-window-size-listener';
-import slug from 'slug';
-import { stringify } from 'qs';
-import { setLocation } from '../utils/common';
+import React, { PureComponent } from "react";
+import { withStyles } from "@material-ui/core/styles";
+import Grid from "@material-ui/core/Grid";
+import Button from "@material-ui/core/Button";
+import AppLayout from "../components/layouts/AppLayout";
+import { Typography } from "@material-ui/core";
+import styles from "../styles/common";
+import classnames from "classnames";
+import Link from "next/link";
+import notify from "../utils/notifier";
+import { userAPI } from "../services/userAPI";
+import Geocode from "react-geocode";
+import { APP_URL, API_IMAGE_URL } from "../utils/config";
+import RoomIcon from "@material-ui/icons/Room";
+import { Scrollbars } from "react-custom-scrollbars";
+import SocialLinks from "../components/common/SocialLinks";
+import actions from "../redux/global/actions";
+import { connect } from "react-redux";
+import WindowResizeListener from "react-window-size-listener";
+import slug from "slug";
+import { stringify } from "qs";
+import { setLocation } from "../utils/common";
 const { setCurrentLocation } = actions;
 
 class WelcomeToDishIn extends PureComponent {
   state = {
-    searchValue: '',
+    searchValue: "",
     restaurants: [],
     enableApi: true,
     selectedIndex: -1,
-    restaurantName: '',
-    winHeight: '100vh'
+    restaurantName: "",
+    winHeight: "100vh",
   };
   static getInitialProps({ store, isServer }) {
     return { isServer };
@@ -38,22 +37,22 @@ class WelcomeToDishIn extends PureComponent {
   componentDidMount() {
     if (navigator.geolocation) {
       navigator.geolocation.getCurrentPosition(
-        position => {
+        (position) => {
           var pos = {
             lat: position.coords.latitude,
-            lng: position.coords.longitude
+            lng: position.coords.longitude,
           };
           // set Google Maps Geocoding API for purposes of quota management. Its optional but recommended.
-          Geocode.setApiKey('AIzaSyChP5Ri3hwQG4BRFzmDxqGE_SHQnJwPkjc');
+          Geocode.setApiKey("AIzaSyChP5Ri3hwQG4BRFzmDxqGE_SHQnJwPkjc");
           // Enable or disable logs. Its optional.
           Geocode.enableDebug();
 
           // Get address from latidude & longitude.
           Geocode.fromLatLng(pos.lat, pos.lng).then(
-            response => {
-              let address = '';
-              const streetAddress = response.results.find(el => {
-                return el.types.includes('street_address');
+            (response) => {
+              let address = "";
+              const streetAddress = response.results.find((el) => {
+                return el.types.includes("street_address");
               });
               if (streetAddress) {
                 address = streetAddress.formatted_address;
@@ -62,7 +61,7 @@ class WelcomeToDishIn extends PureComponent {
               }
               this.updateLocation({ ...pos, address });
             },
-            error => {
+            (error) => {
               this.updateLocation({ ...pos });
             }
           );
@@ -86,7 +85,7 @@ class WelcomeToDishIn extends PureComponent {
     }
   }
 
-  updateLocation = location => {
+  updateLocation = (location) => {
     const { setCurrentLocation } = this.props;
     setLocation(location);
     setCurrentLocation(location);
@@ -95,27 +94,25 @@ class WelcomeToDishIn extends PureComponent {
     }
   };
 
-  getRestaurantAvatar = rec => {
+  getRestaurantAvatar = (rec) => {
     if (rec.images.length) {
-      return `${API_IMAGE_URL}/assets/images/restaurants/${rec.slug}/${
-        rec.images[0].path
-      }`;
+      return `${API_IMAGE_URL}/assets/images/restaurants/${rec.slug}/${rec.images[0].path}`;
     } else {
-      return '';
+      return "";
     }
   };
 
-  loadRestaurants = location => {
+  loadRestaurants = (location) => {
     location = location || {};
     userAPI
-      .getRestaurants({ name: '', location: { ...location } })
-      .then(response => {
-        if (response.status === 'ok') {
-          let restaurants = response.data.map(rec => {
+      .getRestaurants({ name: "", location: { ...location } })
+      .then((response) => {
+        if (response.status === "ok") {
+          let restaurants = response.data.map((rec) => {
             return {
               avatar: this.getRestaurantAvatar(rec.restaurant_id),
               label: rec.restaurant_id.name,
-              id: rec.restaurant_id._id
+              id: rec.restaurant_id._id,
             };
           });
           this.setState({ restaurants });
@@ -132,42 +129,41 @@ class WelcomeToDishIn extends PureComponent {
 
   handListItemClick = (evt, selectedIndex) => {
     evt.preventDefault();
-    const found = this.state.restaurants[selectedIndex] || { primary: '' };
+    const found = this.state.restaurants[selectedIndex] || { primary: "" };
     if (found.primary) {
       this.setState({ selectedIndex, restaurantName: found.primary });
     } else {
-      this.setState({ selectedIndex: '', restaurantName: '' });
+      this.setState({ selectedIndex: "", restaurantName: "" });
     }
   };
 
   renderThumb = ({ style, ...props }) => {
     const thumbStyle = {
-      backgroundColor: 'rgba(240,242,245,.5)',
-      border: '1px solid rgba(0,0,0,.3)'
+      backgroundColor: "rgba(240,242,245,.5)",
+      border: "1px solid rgba(0,0,0,.3)",
     };
     return <div style={{ ...style, ...thumbStyle }} {...props} />;
   };
 
-  handleSubmit = evt => {
+  handleSubmit = (evt) => {
     evt.preventDefault();
     const selectedItem = this.autoComplete.selectedItem;
     if (selectedItem && selectedItem.selectedItem) {
-      window.location.href = `/restaurants/${slug(
-        selectedItem.selectedItem,
-        { lower: true }
-      )}`;
+      window.location.href = `/restaurants/${slug(selectedItem.selectedItem, {
+        lower: true,
+      })}`;
     } else {
       notify(`Please select a restaurant`);
     }
   };
-  autoCompleteRef = ref => {
+  autoCompleteRef = (ref) => {
     this.autoComplete = ref;
   };
 
   render() {
     const {
       classes,
-      global: { location }
+      global: { location },
     } = this.props;
     const { restaurants, winHeight } = this.state;
     let adjustHeight = 5;
@@ -183,41 +179,41 @@ class WelcomeToDishIn extends PureComponent {
     return (
       <AppLayout {...this.props}>
         <WindowResizeListener
-          onResize={windowSize => {
+          onResize={(windowSize) => {
             this.setState({ winHeight: windowSize.windowHeight });
           }}
         />
-        <form className={classes.container} noValidate autoComplete='off'>
+        <form className={classes.container} noValidate autoComplete="off">
           <Grid
             container
-            direction='column'
-            justify='space-between'
-            alignItems='center'
+            direction="column"
+            justify="space-between"
+            alignItems="center"
             spacing={0}
             style={{
-              margin: '0px 16px',
-              minHeight: `${rootHeight}px`
+              margin: "0px 16px",
+              minHeight: `${rootHeight}px`,
             }}
           >
             <Grid
               item
               style={{
-                backgroundColor: '#f0f0f0',
+                backgroundColor: "#f0f0f0",
                 height: `${adjustHeight}px`,
-                width: '100%'
+                width: "100%",
               }}
             />
             <Grid
               item
               style={{
-                backgroundColor: '#666',
-                height: '60px',
-                width: '100%'
+                backgroundColor: "#666",
+                height: "60px",
+                width: "100%",
               }}
             >
               <Typography
-                variant='h1'
-                align='center'
+                variant="h1"
+                align="center"
                 className={classes.pageTitleRed}
               >
                 WELCOME TO DISHIN
@@ -225,9 +221,9 @@ class WelcomeToDishIn extends PureComponent {
               <div
                 className={classes.footerLatoTextNormal}
                 style={{
-                  margin: '0 26px',
-                  marginTop: '10px',
-                  textAlign: 'center'
+                  margin: "0 26px",
+                  marginTop: "10px",
+                  textAlign: "center",
                 }}
               >
                 {location.address ? (
@@ -236,32 +232,32 @@ class WelcomeToDishIn extends PureComponent {
                     {location.address}
                   </React.Fragment>
                 ) : (
-                  'Where are you dining today?'
+                  "Where are you dining today?"
                 )}
               </div>
             </Grid>
             <Grid
               item
               style={{
-                backgroundColor: '#f0f0f0',
+                backgroundColor: "#f0f0f0",
                 height: `${adjustHeight}px`,
-                width: '100%'
+                width: "100%",
               }}
             />
             <Grid
               item
               style={{
-                backgroundColor: '#ddd',
-                height: '150px',
-                width: '100%'
+                backgroundColor: "#ddd",
+                height: "150px",
+                width: "100%",
               }}
             >
-              <div style={{ margin: '0 26px' }}>
+              <div style={{ margin: "0 26px" }}>
                 <AutoComplete
                   innerRef={this.autoCompleteRef}
-                  id='restaurantName'
-                  name='restaurantName'
-                  placeholder='Restaurant Name'
+                  id="restaurantName"
+                  name="restaurantName"
+                  placeholder="Restaurant Name"
                   data={restaurants}
                   isOpen={restaurants.length ? true : false}
                 />
@@ -270,22 +266,22 @@ class WelcomeToDishIn extends PureComponent {
             <Grid
               item
               style={{
-                backgroundColor: '#f0f0f0',
+                backgroundColor: "#f0f0f0",
                 height: `${adjustHeight}px`,
-                width: '100%'
+                width: "100%",
               }}
             />
             <Grid
               item
               style={{
-                backgroundColor: '#dadada',
-                height: '36px',
-                width: '100%'
+                backgroundColor: "#dadada",
+                height: "36px",
+                width: "100%",
               }}
             >
-              <div style={{ margin: '0 40px' }}>
+              <div style={{ margin: "0 40px" }}>
                 <Button
-                  size='medium'
+                  size="medium"
                   className={classes.btnRaisedLightNormalRed}
                   fullWidth
                   onClick={this.handleSubmit}
@@ -297,23 +293,23 @@ class WelcomeToDishIn extends PureComponent {
             <Grid
               item
               style={{
-                backgroundColor: '#f0f0f0',
+                backgroundColor: "#f0f0f0",
                 height: `${abc}px`,
-                width: '100%'
+                width: "100%",
               }}
             />
             <Grid
               item
               style={{
-                backgroundColor: '#fafafa',
-                height: '197px',
-                width: '100%',
-                textAlign: 'center'
+                backgroundColor: "#fafafa",
+                height: "197px",
+                width: "100%",
+                textAlign: "center",
               }}
             >
               <Typography>&#160;</Typography>
               <div className={classes.footerLatoTextNormal}>
-                New User?{' '}
+                New User?{" "}
                 <a
                   href={`/sign-up`}
                   className={classnames(
@@ -322,7 +318,7 @@ class WelcomeToDishIn extends PureComponent {
                   )}
                 >
                   Sign Up
-                </a>{' '}
+                </a>{" "}
                 to the app
               </div>
               <Typography>&#160;</Typography>
@@ -332,7 +328,7 @@ class WelcomeToDishIn extends PureComponent {
               <Typography>&#160;</Typography>
               <SocialLinks />
               <div className={classes.footerLatoTextNormal}>
-                Forgot Password?{' '}
+                Forgot Password?{" "}
                 <a
                   href={`/recover-password`}
                   className={classnames(
@@ -352,10 +348,10 @@ class WelcomeToDishIn extends PureComponent {
 }
 
 export default connect(
-  state => ({
-    global: state.global
+  (state) => ({
+    global: state.global,
   }),
   {
-    setCurrentLocation
+    setCurrentLocation,
   }
 )(withStyles(styles)(WelcomeToDishIn));
