@@ -1,10 +1,19 @@
-import { LIST_DISHES } from "./types";
-const intialState = [];
+import { GET_DISHES, SEARCH_DISHES, CLEAR } from "./types";
+
+const intialState = { dishes: [] };
 
 export default function DishesReducer(state = intialState, action) {
   const { type, payload } = action;
   switch (type) {
-    case LIST_DISHES:
+    case GET_DISHES:
+      console.log("payload", payload);
+      return {
+        ...state,
+        dishes: payload,
+      };
+    case SEARCH_DISHES:
+      return payload;
+    case CLEAR:
       return payload;
 
     default:
