@@ -1,5 +1,5 @@
 import { userReducer as user } from "./user/reducer";
 import { globalReducer as global } from "./global/reducer";
 import RestaurantsReducer from "./restaurants/reducer";
-import { DishesReducer } from "./dishes/reducers";
-export { user, global, RestaurantsReducer };
+import DishesReducer from "./dishes/reducers";
+export { user, global, RestaurantsReducer, DishesReducer };
