@@ -1,4 +1,4 @@
-import { openSnackbar } from '../components/common/Notifier';
+import { openSnackbar } from "../components/Notifier/Notifier";
 
 export default function notify(obj) {
   openSnackbar({ message: obj.message || obj.toString() });
